requirements.txt
moviepy==1.0.3
fpdf==1.7.2
Pillow==10.3.0